Do not use config to store the infra app aliases
