+++ resotocore/resotocore/__main__.py
@@ -184,7 +184,9 @@ def with_config(
     cli_deps.extend(task_handler=task_handler, inspector=inspector)
     infra_apps_runtime = LocalResotocoreAppRuntime(cli)
     cli_deps.extend(infra_apps_runtime=infra_apps_runtime)
-    infra_apps_package_manager = PackageManager(db.package_entity_db, config_handler)
+    infra_apps_package_manager = PackageManager(
+        db.package_entity_db, config_handler, cli.register_alias_template, cli.unregister_alias_template
+    )
     cli_deps.extend(infra_apps_package_manager=infra_apps_package_manager)
     graph_manager = GraphManager(db, config.snapshots, core_config_handler, task_handler)
     cli_deps.extend(graph_manager=graph_manager)
