+++ resotocore/resotocore/web/api.py
@@ -62,10 +62,10 @@ from resotocore.cli.model import (
     CLIContext,
     CLICommand,
     InternalPart,
-    AliasTemplate,
     WorkerCustomCommand,
     CLI,
 )
+from resotocore.cli.alias_template import AliasTemplate
 from resotocore.config import ConfigHandler, ConfigValidation, ConfigEntity
 from resotocore.console_renderer import ConsoleColorSystem, ConsoleRenderer
 from resotocore.core_config import CoreConfig
@@ -731,7 +731,7 @@ class Api:
             worker_descriptions.set_result(description)
             # register the descriptions as custom command on the CLI
             for cmd in cmds:
-                self.cli.register_worker_custom_command(cmd)
+                self.cli.register_alias_template(cmd.to_template())
             # the connect process is done, define the final handler
             handler = handle_message
