+++ resotocore/tests/resotocore/conftest.py
@@ -533,7 +533,7 @@ async def inspector_service(cli: CLIService) -> InspectorService:
 async def package_manager(
     cli: CLIService, config_handler: ConfigHandler, package_entity_db: PackageEntityDb
 ) -> PackageManager:
-    async with PackageManager(package_entity_db, config_handler) as service:
+    async with PackageManager(package_entity_db, config_handler, lambda at: None, lambda s: None) as service:
         cli.dependencies.lookup["infra_apps_package_manager"] = service
         return service
